Avoid secret collisions
This patch adds support for the `kubernetes.io/tls-acme-secret-name`
annotation. When set, the secret used to store the TLS certificate will
be named after its value.

If unset, the name of the route will be used instead.

Example usage:

    oc annotate route/myapp \
        kubernetes.io/tls-acme="true" \
        kubernetes.io/tls-acme-secret-name="myapp-tls-cert"

This will obtain a TLS certificate and store it in a secret called
`myapp-tls-cert`.

Fixes: #64
Signed-off-by: Manuel Hutter <manuel@vshn.ch>